fix: fix color sensor reflection and init turret angle

+++ TeamCode/src/main/java/sigmacorns/io/SigmaIO.kt
@@ -42,7 +42,7 @@ interface SigmaIO {
     fun spindexerPosition(): Double
     fun distance(): Double
 
-    fun setTurretPosition(Offset: Int)
+    fun setTurretPosition(Offset: Double)
     fun update()
 
     fun setPosition(p: Pose2d)
